pyproject.toml
[tool.poetry]
name = "ai-agents-react"
version = "0.1.0"
description = ""
authors = ["Samuel Matioli <samuel.matioli@datastax.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
langchain = "^0.2.15"
astrapy = "^1.4.1"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.14"
openai = "^1.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
